systemd: add gpg agent socket overrides for smartcard.target
With these modifications, we can simply enable the sockets and the
gpg-agent will come up when smartcard.target is reached when a
suitable card is plugged in. Removing it will stop the sockets and
remove the associated files, so no problems are caused when logging
into a given system and forwarding the sockets from elsewhere.

Signed-off-by: John Helmert III <jchelmert3@posteo.net>